Don't leak Pango::Attribute objects
The various Pango::Attribute subclasses, like Pango::AttrSize, are not
registered with the Glib::Boxed type machinery.  So when a variable
holding one of these goes out of scope, Glib::Boxed::DESTROY tries to
look up the calling class name, say "Pango::AttrSize", in its registry
so that it can call the right function to reclaim the memory, but it
finds nothing.  To make this work, register all the Perl subclasses as
aliases of PANGO_TYPE_ATTRIBUTE, using gperl_register_boxed_alias.

https://bugzilla.gnome.org/show_bug.cgi?id=616894